+++ src/test/java/com/stcom/smartmealtable/domain/member/MemberPasswordTest.java
@@ -5,6 +5,8 @@ import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
+import com.stcom.smartmealtable.exception.PasswordFailedExceededException;
+import com.stcom.smartmealtable.exception.PasswordPolicyException;
 import org.junit.jupiter.api.Test;
 
 class MemberPasswordTest {
